fix: views permissions

+++ lgu/views.py
@@ -7,14 +7,14 @@ from registrant.forms import AddressFieldForm
 
 
 def DashboardView(request, lgu_id):
-    # if not request.user.is_lgu:
-    #     return render(request, "allow.html", {})
+    if not request.user.is_lgu:
+        return render(request, "allow.html", {})
     obj = LocalGovernmentUnit.objects.get(pk=lgu_id)
     first_dose = obj.individuals.filter(vaccination_status=1).count()
     fully_vaccinated = obj.individuals.filter(vaccination_status=2).count()
     template = "lguDashboard.html"
-    # if request.user.id != obj.user.id:
-    #     template = "allow.html"
+    if request.user.id != obj.user.id:
+        template = "allow.html"
     generate_registrant_markers_map(obj)
     context = {
         'obj': obj,
@@ -39,8 +39,8 @@ def VaccinationSiteView(request, vs_id, lgu_id):
 
 
 def VaccinationSiteCreate(request, lgu_id):
-    # if not request.user.is_lgu:
-    #     return render(request, "allow.html", {})
+    if not request.user.is_lgu:
+        return render(request, "allow.html", {})
     site_form = VaccinationSiteForm(request.POST or None)
     address_form = AddressFieldForm(request.POST or None)
     lgu = LocalGovernmentUnit.objects.get(pk=lgu_id)
@@ -62,8 +62,8 @@ def VaccinationSiteCreate(request, lgu_id):
     return render(request, "vaccinationSite.html", context)
 
 def VaccinationSiteUpdate(request, lgu_id, vs_id):
-    # if not request.user.is_lgu:
-    #     return render(request, "allow.html", {})
+    if not request.user.is_lgu:
+        return render(request, "allow.html", {})
     vs = VaccinationSite.objects.get(pk=vs_id)
     address = vs.address
     site_form = VaccinationSiteForm(request.POST or None, instance=vs)
@@ -88,8 +88,8 @@ def VaccinationSiteUpdate(request, lgu_id, vs_id):
 
 
 def VaccinationSiteDelete(request, lgu_id, vs_id):
-    # if not request.user.is_lgu:
-    #     return render(request, "allow.html", {})
+    if not request.user.is_lgu:
+        return render(request, "allow.html", {})
     obj = VaccinationSite.objects.get(id=vs_id)
     if request.method == 'POST':
         obj.delete()
@@ -102,8 +102,8 @@ def VaccinationSiteDelete(request, lgu_id, vs_id):
 
 
 def PriorityLocationCreate(request, lgu_id):
-    # if not request.user.is_lgu:
-    #     return render(request, "allow.html", {})
+    if not request.user.is_lgu:
+        return render(request, "allow.html", {})
     priority_form = PriorityLocationForm(request.POST or None)
     address_form = AddressFieldForm(request.POST or None)
     lgu = LocalGovernmentUnit.objects.get(pk=lgu_id)
@@ -128,8 +128,8 @@ def PriorityLocationCreate(request, lgu_id):
 
 
 def PriorityLocationUpdate(request, lgu_id, pl_id):
-    # if not request.user.is_lgu:
-    #     return render(request, "allow.html", {})
+    if not request.user.is_lgu:
+        return render(request, "allow.html", {})
     pl = PriorityLocation.objects.get(pk=pl_id)
     address = pl.address
     priority_form = PriorityLocationForm(request.POST or None, instance=pl)
@@ -156,8 +156,8 @@ def PriorityLocationUpdate(request, lgu_id, pl_id):
 
 
 def PriorityLocationDelete(request, lgu_id, pl_id):
-    # if not request.user.is_lgu:
-    #     return render(request, "allow.html", {})
+    if not request.user.is_lgu:
+        return render(request, "allow.html", {})
     obj = PriorityLocation.objects.get(id=pl_id)
     if request.method == 'POST':
         obj.delete()
